requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-filter==1.1.0
django-simple-captcha==0.5.3
djangorestframework==3.7.3
djangorestframework-jwt==1.9.0
idna==2.6
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
six==1.11.0
testfixtures==5.3.1
urllib3==1.22
yunpian-python-sdk==1.0.0
